requirements.txt
# Dépendances Python pour le projet
pandas>=1.5.0
requests>=2.28.0
pyarrow>=10.0.0
numpy>=1.21.0
apache-airflow==2.7.1

# Optionnel pour les visualisations
matplotlib>=3.5.0
seaborn>=0.11.0

# Optionnel pour les notifications
slack-sdk>=3.0.0
